index full text changes
