+++ crates/runtime/src/embeddings/connector.rs
@@ -14,6 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 use crate::accelerated_table::AcceleratedTable;
+use crate::changes::index_change_envelope;
 use crate::component::ComponentInitialization;
 use crate::component::dataset::Dataset;
 use crate::component::metrics::MetricsProvider;
@@ -249,31 +250,6 @@ impl EmbeddingConnector {
         }
     }
 
-    async fn index_change_envelope(
-        maybe_envelope: Result<ChangeEnvelope, StreamError>,
-        embedding_table: Arc<IndexedTableProvider>,
-    ) -> Result<ChangeEnvelope, StreamError> {
-        let envelope = maybe_envelope.map_err(|e| {
-            tracing::debug!("Error in underlying base stream: {e:?}");
-            e
-        })?;
-
-        let (change_committer, batch) = envelope.into_parts();
-        let mut batches = vec![batch.data_batch()];
-
-        for index in &embedding_table.indexes {
-            batches = index
-                .compute_index(batches)
-                .await
-                .map_err(|e| StreamError::External(e.to_string()))?;
-        }
-
-        let new_change_batch = replace_change_batch_data(&batches[0], &batch)
-            .map_err(|e| StreamError::Arrow(e.to_string()))?;
-
-        Ok(ChangeEnvelope::new(change_committer, new_change_batch))
-    }
-
     async fn embed_change_envelope(
         maybe_envelope: Result<ChangeEnvelope, StreamError>,
         embedding_table: Arc<EmbeddingTable>,
@@ -387,7 +363,7 @@ impl DataConnector for EmbeddingConnector {
             let stream = self
                 .inner_connector
                 .changes_stream(underlying_federated_table)?
-                .then(move |item| Self::index_change_envelope(item, Arc::clone(&indexed_table)))
+                .then(move |item| index_change_envelope(item, Arc::clone(&indexed_table)))
                 .boxed();
 
             return Some(stream);
@@ -430,7 +406,7 @@ impl DataConnector for EmbeddingConnector {
             let stream = self
                 .inner_connector
                 .append_stream(underlying_federated_table)?
-                .then(move |item| Self::index_change_envelope(item, Arc::clone(&indexed_table)))
+                .then(move |item| index_change_envelope(item, Arc::clone(&indexed_table)))
                 .boxed();
 
             return Some(stream);
